4.2 - 🦀 Cleanup of tokens.rs
The to_string() is not how we do it in Rust, instead we implement the
Display trait if we want to format the public description of the object.

I also renamed all of the items in the TokenType enum, as both rust_fmt
and clippy was more then reasonably upset with the code. In Rust the
standard is to use CamelCase for Types, and snake_case for functions,
variables and parameters.