4.4 - The Scanner Class
So, I had already started on this, just to get Rust happy, but now we
catch up to this in the book, and start filling it out with some logic.
To get this code to work, I had to declare the scanner instance as mut,
as it needs to mutate itself (update internal fields).

This scanner could probably be some variant of an iterator in Rust, but
yet again, let us instead try to follow the example code as closely as
possible.