4.5.1 - Lexical errors
We start to see the how the code works with the error reporting here, but
we will push on a bit more until we Rustify it, so we don't reach any pre-
mature conclusions. But yes, it starts to look like we could introduce
the normal Rust Result<T,E> type here. :)