4.6 - 🦀 Cleanup of the number() function
First, our own function is_digit() was removed in favor of the
char::is_ascii_digit(), as my IDE was complaining about it, and
recommended me to switch to it.

This change helped me refactor the implementation of number() to
not rely on a local variable to check the different conditions, which made
it much nicer. :)