4.7 - 🦀 Some cleanup
The advance() method returned the current char and increased the
current counter, but we only used the return value in 1 of 10 places.
So I made that method only responsible for increasing the counter, and
added an extra call to peek() where we used the value.

A small cleaning up of the is_alpha() and is_alpha_numeric() methods,
implementing them with the help of the built in functions on char.

Added next_line() to increase the line counter.

Some other small changes, formatting and such.