4.7 - 🦀 Fix add_token_object()
We're finally at the the time to fix the old add_token_object() method.

The Token.literal field was originally typed as an Object in the Java
code, but that doesn't fit well into Rust's stricter type system.

Our Literal type is an Enum with two variants, either A String(String) or
a Number(f32), which means that we now have the possibility to store
either String-literals or Number-literals along with their values. 👍